pyproject.toml
[project]
name = "artbox"
version = "0.10.0"  # semantic-release
description = "ArtBox is a tool set for handling multimedia files."
readme = "docs/index.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.9,<3.14"
authors = [
  {name = "Ivan Ogasawara", email = "ivan.ogasawara@gmail.com"},
]
dependencies = [
  "pycairo>=1.25.1,<1.26.0",
  "pygobject>=3.44.1,<3.49",
  "openai>=1",
  "python-dotenv>=1.0.0",
  "moviepy>=1.0.3,<2",
  "pydub>=0.25.1",
  "noisereduce>=2.0.1,<3",
  "aubio>=0.4.9",
  "gtts>=2.3.2",
  "edge-tts>=6.1.8",
  "numpy>=1.20,<2",
  "typer>=0.9.0",
  "pytubefix>=5.0",
  "speechrecognition>=3.10",
  "vosk>=0.3.45",
  "google-cloud-speech>=2.24.1",
  "matplotlib<=3.9",
  "scipy<1.23",
  "librosa>=0.10.1",
  "pyyaml>=6.0",
  "jsonschema>=4.0",
  "pdf2image>=1.16",
  "ffmpeg-python>=0.2.0",
  "types-tqdm (>=4.67.3.20260205,<5.0.0.0)",
]

[project.scripts]
artbox = "artbox.__main__:app"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
  {include = "artbox", from = "src"},
]
include = ["src/artbox/*.json"]
exclude = [
  ".git/*",
  ".env*",
]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.2"
pytest-cov = ">=4.1.0"
coverage = ">=7.2.7"
pre-commit = ">=3.3.2"
ruff = ">=0.1.5"
mypy = ">=1.6.0"
bandit = ">=1.7.5"
vulture = ">=2.9.1"
mccabe = ">=0.6.1"
ipykernel = ">=6.0.0"
ipython = ">=7.0.0"
mkdocs = ">=1.3"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.24.7"
mkdocs-literate-nav = ">=0.4.1"
mkdocs-macros-plugin = ">=0.6.3"
mkdocs-material = ">=8"
mkdocstrings = {version=">=0.19.0", extras=["python"]}
mkdocstrings-python = "<2"
jupyterlab = ">=4.0.5"
makim = "1.13.0"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.vulture]
exclude = ["tests"]
ignore_decorators = []
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["src/artbox"]
sort_by_size = true
verbose = false

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
fix = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "PL",  # PL
  "RUF", # Ruff-specific rules
  "I001", # isort
]
ignore = ["PLR0913", "PLR0915"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.ruff.format]
quote-style = "double"


[tool.mypy]
no_strict_optional = false

[[tool.mypy.overrides]]
module = [
  "aubio",
  "edge_tts",
  "gtts",
  "moviepy.editor",
  "noisereduce",
  "pydub",
  "pydub.generators",
  "pytubefix",
  "speech_recognition",
  "ffmpeg",
  "matplotlib.pyplot",
  "matplotlib",
  "yaml",
  "jsonschema",
  "pdf2image",
]
ignore_missing_imports = true
